jankend/requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-SocketIO==5.3.0
Flask-Cors==3.1.1
python-dotenv==1.0.0
paho-mqtt==1.6.1
